Update README.md
Abstract:

This project presents the development of an Internet of Things (IoT) based environmental monitoring system using Arduino. The system is designed to monitor and collect data on environmental parameters such as temperature and humidity. It leverages the Arduino microcontroller, DHT22 sensor, Wi-Fi connectivity, and MQTT (Message Queuing Telemetry Transport) protocol for data transmission to a cloud-based server.

The hardware setup involves connecting a DHT22 sensor to the Arduino, which measures temperature and humidity levels in real-time. The Arduino is configured to establish a Wi-Fi connection to the local network, and it communicates with an MQTT broker. The sensor data is then published to specific MQTT topics.

On the cloud side, an MQTT broker is deployed to receive and store the sensor data. Users can subscribe to these topics to access real-time environmental information. The data can be further analyzed, visualized, or used to trigger specific actions, making it a versatile tool for environmental monitoring and automation.

This IoT-based environmental monitoring system offers scalability, flexibility, and real-time data accessibility, making it suitable for a wide range of applications, from home automation to industrial environments and agriculture. It contributes to the growing field of IoT, enabling users to make informed decisions based on real-time environmental data